requirements.txt
argopy
xarray
netCDF4
pandas 
numpy
pyarrow
sqlalchemy 
streamlit
plotly
langchain
langchain-community
langchain-groq
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
pydantic
python-dotenv